capi: refactor module
